requirements.txt
streamlit>=1.52.0
pandas>=2.3.0
python-dotenv>=1.0.0
requests>=2.32.0
gspread>=6.2.0
google-auth>=2.40.0
pytest>=8.3.0
